+++ NEWS.md
@@ -29,12 +29,13 @@
 
 * `char_mat()` and `num_mat()` take lists and convert them to binary, value, or
   count matrices.
+* `cSplit_cartesian()` splits data into a long form returning the Cartesian 
+  Product of the split values.
+* `move_me()` moves values around using natural language statements.
 * `trim_vec()` and `trim_list()` take vectors or lists and trims the whitespace,
   being careful about empty elements and zero-length elements.
 * `unlist_cols()` added as a convenience function for `listCol_l()` and 
   `listCol_w()`.
-* `cSplit_cartesian()` splits data into a long form returning the Cartesian 
-  Product of the split values.
   
 ### Enhancements